Dumbo Completed

+++ db/migrate/005_create_companies.rb
@@ -3,7 +3,7 @@ class CreateCompanies < ActiveRecord::Migration[4.2]
   def change
     create_table :companies do |t|
       t.string :name
-      t.integer :balance
+      t.float :balance
     end
   end
 end
